requirements.txt
requests
boto3
kafka-python
psycopg[binary]
pandas
pytest
apache-airflow==2.8.1